requirements.txt
reportlab
babel
py-dateutil
vobject
caldav
configobj
cherrypy